requirements.txt
# ============================================
# Movie Mood Recommender - Dependencies
# ============================================

# Data Collection
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
tqdm>=4.66.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# NLP & ML (prep for Phase 2)
transformers>=4.36.0
torch>=2.1.0
nltk>=3.8.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# Development
jupyter>=1.0.0
ipykernel>=6.27.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Optional: For async scraping
# aiohttp>=3.9.0
# asyncio>=3.4.3